+++ client-react/src/pages/app/deployment-center/github-provider/DeploymentCenterGitHubAccount.tsx
@@ -1,63 +1,24 @@
 import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { DeploymentCenterGitHubProviderProps } from '../DeploymentCenter.types';
-import { PrimaryButton, Label, Link, IDropdownOption, MessageBarType } from 'office-ui-fabric-react';
+import { PrimaryButton, Label, Link, MessageBarType } from 'office-ui-fabric-react';
 import ReactiveFormControl from '../../../../components/form-controls/ReactiveFormControl';
 import { additionalTextFieldControl, deploymentCenterInfoBannerDiv } from '../DeploymentCenter.styles';
-import Dropdown from '../../../../components/form-controls/DropDown';
-import { Field } from 'formik';
 import CustomBanner from '../../../../components/CustomBanner/CustomBanner';
 import { DeploymentCenterLinks } from '../../../../utils/FwLinks';
 
 const DeploymentCenterGitHubAccount: React.FC<DeploymentCenterGitHubProviderProps> = props => {
   const {
-    formProps,
     accountUser,
     accountStatusMessage,
     authorizeAccount,
-    fetchRepositoryOptions,
-    fetchBranchOptions,
-    organizationOptions,
-    repositoryOptions,
-    branchOptions,
   } = props;
   const { t } = useTranslation();
 
   const [showInfoBanner, setShowInfoBanner] = useState(true);
-  const [selectedOrg, setSelectedOrg] = useState<string>('');
-  const [selectedRepo, setSelectedRepo] = useState<string>('');
-  const [selectedBranch, setSelectedBranch] = useState<string>('');
 
   const closeInfoBanner = () => {
     setShowInfoBanner(false);
-  };
-
-  const onOrganizationChange = (event: React.FormEvent<HTMLDivElement>, option: IDropdownOption) => {
-    setSelectedOrg(option.key.toString());
-    formProps.setFieldValue('org', option.text);
-
-    setSelectedRepo('');
-    formProps.setFieldValue('repo', '');
-
-    setSelectedBranch('');
-    formProps.setFieldValue('branch', '');
-
-    fetchRepositoryOptions(option.key.toString());
-  };
-
-  const onRepositoryChange = (event: React.FormEvent<HTMLDivElement>, option: IDropdownOption) => {
-    setSelectedRepo(option.text);
-    formProps.setFieldValue('repo', option.key.toString());
-
-    setSelectedBranch('');
-    formProps.setFieldValue('branch', '');
-
-    fetchBranchOptions(formProps.values.org, option.text);
-  };
-
-  const onBranchChange = async (event: React.FormEvent<HTMLDivElement>, option: IDropdownOption) => {
-    setSelectedBranch(option.text);
-    formProps.setFieldValue('branch', option.key.toString());
   };
 
   const gitHubAccountControls = accountUser ? (
@@ -85,48 +46,12 @@ const DeploymentCenterGitHubAccount: React.FC<DeploymentCenterGitHubProviderProp
           </Link>
         </div>
       </ReactiveFormControl>
-      <Field
-        id="deployment-center-settings-organization-option"
-        label={t('deploymentCenterOAuthOrganization')}
-        placeholder={t('deploymentCenterOAuthOrganizationPlaceholder')}
-        name="org"
-        component={Dropdown}
-        displayInVerticalLayout={true}
-        options={organizationOptions}
-        selectedKey={selectedOrg}
-        required={true}
-        onChange={onOrganizationChange}
-      />
-      <Field
-        id="deployment-center-settings-repository-option"
-        label={t('deploymentCenterOAuthRepository')}
-        placeholder={t('deploymentCenterOAuthRepositoryPlaceholder')}
-        name="repo"
-        component={Dropdown}
-        displayInVerticalLayout={true}
-        options={repositoryOptions}
-        selectedKey={selectedRepo}
-        required={true}
-        onChange={onRepositoryChange}
-      />
-      <Field
-        id="deployment-center-settings-branch-option"
-        label={t('deploymentCenterOAuthBranch')}
-        placeholder={t('deploymentCenterOAuthBranchPlaceholder')}
-        name="branch"
-        component={Dropdown}
-        displayInVerticalLayout={true}
-        options={branchOptions}
-        selectedKey={selectedBranch}
-        required={true}
-        onChange={onBranchChange}
-      />
     </>
   ) : (
-    <PrimaryButton ariaDescription={t('deploymentCenterOAuthAuthorizeAriaLabel')} onClick={authorizeAccount}>
-      {t('deploymentCenterOAuthAuthorize')}
-    </PrimaryButton>
-  );
+      <PrimaryButton ariaDescription={t('deploymentCenterOAuthAuthorizeAriaLabel')} onClick={authorizeAccount}>
+        {t('deploymentCenterOAuthAuthorize')}
+      </PrimaryButton>
+    );
 
   const accountStatusMessageControl = <Label>{accountStatusMessage}</Label>;
